# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/login": {
      "post": {
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "name of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": { "type": "string" }
                },
                "required": ["password"]
              },
              "example": {
                "password": "donttextmeas"
              }
            }
          }
        },

        "responses": {
          "200": {
            "description": "Successful Response. Access Token and its Type displayed."
          },
          "401":{
            "description": "Invalid Credentials"
          }
        }
      }
    },
    "/signup": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "phone": { "type": "string" },
                  "email": { "type": "string" },
                  "name": { "type": "string" }
                },
                "required": ["username", "email", "password", "name", "phone"]
              },
              "example": {
                "username": "thisisit2",
                "password": "donttextmeas",
                "phone": "1231231232",
                "email": "mailername@maildomain.mail",
                "name": "Yourname"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}